+++ platform-tracing-api/src/main/java/space/br1440/platform/tracing/api/runtime/state/VersionedState.java
@@ -1,7 +1,0 @@
-package space.br1440.platform.tracing.api.runtime.state;
-
-public interface VersionedState {
-
-    long version();
-
-}